refactor: move CIMD files into mcp-oauth/ folder
