+++ api/src/services/mcp-oauth/cimd.ts
@@ -1,8 +1,8 @@
 import { isIP } from 'node:net';
 import { useEnv } from '@directus/env';
-import { useLogger } from '../logger/index.js';
-import { getAxios } from '../request/index.js';
-import { OAuthError, validateRedirectUri } from './mcp-oauth/index.js';
+import { useLogger } from '../../logger/index.js';
+import { getAxios } from '../../request/index.js';
+import { OAuthError, validateRedirectUri } from './index.js';
 
 const MIN_TTL_MS = 300_000; // 5 minutes
 const MAX_TTL_MS = 86_400_000; // 24 hours
